timer: xtensa_sys_timer: clear the interrupt after was triggered and handled
If the interrupt is not cleared, the handler is called a few hundred times
and then it gets stuck.

Note that when an interrupt is enabled, it will try in each clock cycle to
make an interrupt request. That's why one should clear the interrupt
request in the interrupt handler.
That way we can be sure that the interrupt handler function will run
when the event gets triggered once again.

Signed-off-by: Iuliana Prodan <iuliana.prodan@nxp.com>